requirements.txt
transformers
torchvision
open_clip_torch